# MAINT: Improve buffer speed