Tries to partially address gh-16518, but in the end only does 10-15% difference. The deletion of the scalar branch is a follow up to gh-16389 (it was an oversight that I forgot to delete it in the spin-off from the larger change.

A total of 40+% increase over the current speed is easiy (and tidier than the current code) if extend the ndarray struct by an additional pointer to hold the buffer-info.  Although maybe more importantly, that would also speeds up *all* array operations by a small but non-negligible amount.

Further improvements may be possible by trying to speed up the buffer-format in general (although IIRC, at least half of that slowness was the alignment calculation, which this optimizes away).