@mattip 2020-06-08
32 bit windows with full asserts is failing: `numpy/core/src/multiarray/buffer.c:149: _is_natively_aligned_at: Assertion `offset == 0' failed.