@eric-wieser 2020-06-08
Re @mattip's comment, it looks like the caller is updating `*offset += descr->elsize;` before checking alignment, when it should be doing it after.