@seberg 2020-06-08
Seems I may have had a slight mixup with the diffs/rebase here :/. In anyway fixed, speedup, seems to be around 20% now (500ns to 400ns).