@vrakesh 2020-06-09
@seberg 

Tested the performance on Mac,

I am getting much  better results for some reason, not sure why (40-50%)

```python
#With This change

In [1]: import numpy as np                                                                                      

In [2]: z = np.zeros(1000)                                                                                      

In [3]: %timeit memoryview(z)                                                                                   
293 ns ± 0.791 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [4]: %timeit memoryview(z)                                                                                   
296 ns ± 3.99 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

```python
#Without change

In [1]: import numpy as np                                                                                      

In [2]: z = np.zeros(1000)                                                                                      

In [3]: %timeit memoryview(z)                                                                                   
550 ns ± 12 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [4]: %timeit memoryview(z)                                                                                   
564 ns ± 35.1 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [5]:   
```