@mattip 2020-06-18
@Qiyu8 it is fine for performance improvements to rely on existing benchmarks.